Java: Move samples to new package; Add jbang support (#2026)
- Fixes the template of `conf.properties` to not have quotes;
- Moves samples to a sub package to avoid conflict when running `jbang`
- Adds `jbang` support to execute the samples from command line
