+++ java/samples/example.conf.properties
@@ -1,6 +1,6 @@
-client.openai.key: "OPEN_AI_KEY"
-client.openai.organizationid: "OPEN_AI_ORGANIZATION_ID"
+client.openai.key=OPEN_AI_KEY
+client.openai.organizationid=OPEN_AI_ORGANIZATION_ID
 
-client.azureopenai.key: "AZURE_OPEN_AI_KEY"
-client.azureopenai.endpoint: "AZURE_OPEN_AI_ENDPOINT"
-client.azureopenai.deploymentname: "AZURE_OPEN_AI_DEPLOYMENT_NAME"
+client.azureopenai.key=AZURE_OPEN_AI_KEY
+client.azureopenai.endpoint=AZURE_OPEN_AI_ENDPOINT
+client.azureopenai.deploymentname=AZURE_OPEN_AI_DEPLOYMENT_NAME
