+++ java/samples/sample-code/src/main/java/com/microsoft/semantickernel/samples/syntaxexamples/javaspecific/Java_Example01_ResourceLoading.java
